SQLDBs
The projects to connect a Flask App and a SQL Database were added.

There is two projects available:
Initial - You can use this project to make sure the connection
and the creation of the database were correct.
Finished - You can see here the full REST API working with SQL DB.