Private API projects now have a dev and secure project
Now exists two projects to use the private API project.

Development Mode:
You need to put your api key in a variable in order to
be able to use the script. DO NOT share your API KEY.
This project is just for development.

Production Mode(Secure way):
This way your api key is secure because it is not in
your code. Do not forget to set the correct
environment variable in order to be able to use
the script, your api key will exist as long as the
environment varible does.